[CI/Build] Isolate client buffer allocation
