+++ mooncake-store/src/utils.cpp
@@ -1,15 +1,12 @@
 #include "utils.h"
 #include "random.h"
 #include "mmap_arena.h"
-#include "config.h"
 #include "common.h"
-#include "ub_allocator.h"
 
 #include "ascii_string.h"
 #include "bool_parser.h"
 #include "environ.h"
 
-#include <Slab.h>
 #include <gflags/gflags.h>
 #include <glog/logging.h>
 #include <ifaddrs.h>
@@ -31,13 +28,6 @@
 #include <sys/mman.h>
 #include <thread>
 #include <vector>
-#ifdef USE_CUDA
-#include <cuda_runtime.h>
-#endif
-#ifdef USE_INTRA_NVLINK
-#include "gpu_vendor/intra_nvlink.h"
-#endif
-
 // Feature flag to enable/disable arena allocator. Disabled by default so the
 // library does not pre-map a large pool unless the operator opts in via gflag
 // or an explicit MC_MMAP_ARENA_POOL_SIZE environment override.
@@ -48,20 +38,6 @@ DEFINE_bool(use_mmap_arena_allocator, false,
 // override).
 DEFINE_uint64(mmap_arena_pool_size, 8ULL * 1024 * 1024 * 1024,
               "Arena allocator pool size in bytes");
-#ifdef USE_ASCEND_DIRECT
-#include "acl/acl.h"
-#endif
-#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
-#include "ascend_allocator.h"
-#endif
-#if defined(USE_SUNRISE)
-#include "sunrise_allocator.h"
-#endif
-
-#ifdef USE_NOF
-#include "spdk/spdk_wrapper.h"
-#endif
-
 #include <ylt/coro_http/coro_http_client.hpp>
 
 namespace mooncake {
@@ -112,80 +88,6 @@ AutoPortBinder::~AutoPortBinder() {
     if (socket_fd_ >= 0) {
         close(socket_fd_);
     }
-}
-
-#ifdef USE_VRAM_SEGMENT
-tl::expected<void *, std::string> allocate_vram_memory(
-    size_t total_size, const std::string &protocol) {
-    cudaError_t res;
-    int device;
-    void *ptr = nullptr;
-    res = cudaGetDevice(&device);
-    if (res != cudaSuccess) {
-        LOG(ERROR) << "VRAM Segment cudaGetDevice failed.";
-        return tl::make_unexpected("VRAM Segment cudaGetDevice failed.");
-    }
-    if (protocol == "nvlink_intra") {
-#ifdef USE_INTRA_NVLINK
-        ptr = allocateFabricMemory_intra(total_size);
-        return ptr;
-#else
-        LOG(ERROR) << "Protocol nvlink_intra need USE_INTRA_NVLINK=ON. Please "
-                      "rebuild mooncake from source.";
-        return tl::make_unexpected("Protocol not supported");
-#endif
-    }
-    res = cudaMalloc((void **)&ptr, total_size);
-    if (res != cudaSuccess) {
-        LOG(ERROR) << "VRAM Segment cudaMalloc failed.";
-        return tl::make_unexpected("VRAM Segment cudaMalloc failed.");
-    }
-    return ptr;
-}
-#endif
-
-void *allocate_buffer_allocator_memory(size_t total_size,
-                                       const std::string &protocol,
-                                       size_t alignment, bool use_spdk_dma) {
-    const size_t default_alignment = facebook::cachelib::Slab::kSize;
-    // Ensure total_size is a multiple of alignment
-    if (alignment == default_alignment && total_size < alignment) {
-        LOG(ERROR) << "Total size must be at least " << alignment;
-        return nullptr;
-    }
-#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
-    if (protocol == "ascend" || protocol == "ubshmem") {
-        return ascend_allocate_memory(total_size, protocol);
-    }
-#endif
-#if defined(USE_SUNRISE)
-    if (protocol == "sunrise_link") {
-        return sunrise_allocate_memory(
-            total_size, alignment,
-            mooncake::globalConfig().sunrise_use_device_mem);
-    }
-#endif
-#if defined(USE_UB)
-    if (protocol == "ub") {
-        return mooncake::ub_allocate_memory(alignment, total_size);
-    }
-#endif
-#ifdef USE_NOF
-    if (use_spdk_dma && total_size > 0) {
-        return mooncake::SpdkWrapper::GetInstance().Alloc(total_size, alignment,
-                                                          -1);
-    }
-#endif
-#ifdef USE_VRAM_SEGMENT
-    auto ret = allocate_vram_memory(total_size, protocol);
-    if (!ret) {
-        LOG(ERROR) << ret.error();
-        return nullptr;
-    }
-    return *ret;
-#endif
-    // Allocate aligned memory
-    return aligned_alloc(alignment, total_size);
 }
 
 // Global arena instance (lazy initialization)
@@ -569,43 +471,6 @@ void *allocate_buffer_numa_segments(size_t total_size,
             return s;
         }() << "]";
     return ptr;
-}
-
-void free_memory(const std::string &protocol, void *ptr, bool use_spdk_dma) {
-#if defined(USE_ASCEND_DIRECT) || defined(USE_UBSHMEM)
-    if (protocol == "ascend" || protocol == "ubshmem") {
-        return ascend_free_memory(protocol, ptr);
-    }
-#endif
-#if defined(USE_SUNRISE)
-    if (protocol == "sunrise_link") {
-        return sunrise_free_memory(ptr);
-    }
-#endif
-#if defined(USE_UB)
-    if (protocol == "ub") {
-        mooncake::ub_free_memory(ptr);
-        return;
-    }
-#endif
-#ifdef USE_NOF
-    // Mirror allocate_buffer_allocator_memory(): a buffer taken from the SPDK
-    // hugepage pool (spdk_zmalloc) must be released with spdk_free, not glibc
-    // free(), which would abort with "free(): invalid pointer".
-    if (use_spdk_dma) {
-        mooncake::SpdkWrapper::GetInstance().Free(ptr);
-        return;
-    }
-#endif
-#ifdef USE_VRAM_SEGMENT
-#ifdef USE_INTRA_NVLINK
-    freeFabricMemory_intra(ptr);
-#else
-    cudaFree(ptr);
-#endif
-    return;
-#endif
-    free(ptr);
 }
 
 tl::expected<std::string, int> httpGet(const std::string &url) {
